Rewritten gem using new template.
