+++ .gitignore
@@ -9,3 +9,4 @@
 
 # rspec failure tracking
 .rspec_status
+/vendor/*
