+++ .travis.yml
@@ -4,4 +4,4 @@ language: ruby
 cache: bundler
 rvm:
   - 2.5.3
-before_install: gem install bundler -v 1.17.3
+before_install: gem install bundler -v 2.0.2
