+++ Gemfile
@@ -1,3 +1,4 @@
-source 'https://rubygems.org'
+source "https://rubygems.org"
 
+# Specify your gem's dependencies in html-proofer-mailto_awesome.gemspec
 gemspec
