+++ Rakefile
@@ -2,5 +2,4 @@ require "bundler/gem_tasks"
 require "rspec/core/rake_task"
 
 RSpec::Core::RakeTask.new(:spec)
-
 task :default => :spec
